+++ backend/app/helpers/load_json_from_file.py
@@ -1,0 +1,16 @@
+from flask import json
+
+
+def load_json_from_file(filepath):
+    try:
+        with open(filepath, 'r', encoding='utf-8') as f:
+            return json.load(f)
+    except FileNotFoundError:
+        print(f"Error: File not found: {filepath}")  
+        return None
+    except json.JSONDecodeError as e:
+        print(f"Error decoding JSON in file {filepath}: {e}")
+        return None
+    except Exception as e:
+        print(f"Error loading JSON from file {filepath}: {e}")  
+        return None
