+++ backend/app/helpers/load_prompt_from_file.py
@@ -1,0 +1,10 @@
+
+def load_prompt_from_file(filepath):
+    try:
+        with open(filepath, 'r', encoding='utf-8') as f:
+            return f.read()
+    except FileNotFoundError:
+        return None
+    except Exception as e:
+        print(f"Error loading prompt from file: {e}")
+        return None
