+++ backend/output/20250228_140625.txt
@@ -1,0 +1,22 @@
+QUERY:
+which ferry is the most powerful? based on ferries-info
+
+RESPONSE:
+The most powerful ferry is "Merkurius" with an installed power of 1470 kW.
+
+SQL QUERIES:
+
+--- Query 1 ---
+SHOW TABLES
+
+--- Query 2 ---
+DESCRIBE ferries_info
+
+--- Query 3 ---
+SELECT 'fragancia' AS ferry_name, fragancia.installed_power_kw FROM ferries_info \nUNION ALL\nSELECT 'jupiter', jupiter.installed_power_kw FROM ferries_info \nUNION ALL\nSELECT 'merkurius', merkurius.installed_power_kw FROM ferries_info \nUNION ALL\nSELECT 'nina', nina.installed_power_kw FROM ferries_info \nUNION ALL\nSELECT 'yxlan', yxlan.installed_power_kw FROM ferries_info \nUNION ALL\nSELECT 'marie', marie.installed_power_kw FROM ferries_info \nUNION ALL\nSELECT 'capella', capella.installed_power_kw FROM ferries_info \nUNION ALL\nSELECT 'linda', linda.installed_power_kw FROM ferries_info \nUNION ALL\nSELECT 'sedna', sedna.installed_power_kw FROM ferries_info \nUNION ALL\nSELECT 'ebba_brahe', ebba_brahe.installed_power_kw FROM ferries_info\nORDER BY installed_power_kw DESC LIMIT 1
+
+--- Query 4 ---
+SELECT 'fragancia' AS ferry_name, fragancia.installed_power_kw \nFROM ferries_info WHERE fragancia.installed_power_kw IS NOT NULL\nUNION ALL\nSELECT 'jupiter', jupiter.installed_power_kw \nFROM ferries_info WHERE jupiter.installed_power_kw IS NOT NULL\nUNION ALL\nSELECT 'merkurius', merkurius.installed_power_kw \nFROM ferries_info WHERE merkurius.installed_power_kw IS NOT NULL\nUNION ALL\nSELECT 'nina', nina.installed_power_kw \nFROM ferries_info WHERE nina.installed_power_kw IS NOT NULL\nUNION ALL\nSELECT 'yxlan', yxlan.installed_power_kw \nFROM ferries_info WHERE yxlan.installed_power_kw IS NOT NULL\nUNION ALL\nSELECT 'marie', marie.installed_power_kw \nFROM ferries_info WHERE marie.installed_power_kw IS NOT NULL\nUNION ALL\nSELECT 'capella', capella.installed_power_kw \nFROM ferries_info WHERE capella.installed_power_kw IS NOT NULL\nUNION ALL\nSELECT 'linda', linda.installed_power_kw \nFROM ferries_info WHERE linda.installed_power_kw IS NOT NULL\nUNION ALL\nSELECT 'sedna', sedna.installed_power_kw \nFROM ferries_info WHERE sedna.installed_power_kw IS NOT NULL\nUNION ALL\nSELECT 'ebba_brahe', ebba_brahe.installed_power_kw \nFROM ferries_info WHERE ebba_brahe.installed_power_kw IS NOT NULL\nORDER BY installed_power_kw DESC LIMIT 1
+
+--- Query 5 ---
+SELECT 'fragancia' AS ferry_name, fragancia.installed_power_kw FROM ferries_info WHERE fragancia.installed_power_kw IS NOT NULL UNION ALL SELECT 'jupiter', jupiter.installed_power_kw FROM ferries_info WHERE jupiter.installed_power_kw IS NOT NULL UNION ALL SELECT 'merkurius', merkurius.installed_power_kw FROM ferries_info WHERE merkurius.installed_power_kw IS NOT NULL UNION ALL SELECT 'nina', nina.installed_power_kw FROM ferries_info WHERE nina.installed_power_kw IS NOT NULL UNION ALL SELECT 'yxlan', yxlan.installed_power_kw FROM ferries_info WHERE yxlan.installed_power_kw IS NOT NULL UNION ALL SELECT 'marie', marie.installed_power_kw FROM ferries_info WHERE marie.installed_power_kw IS NOT NULL UNION ALL SELECT 'capella', capella.installed_power_kw FROM ferries_info WHERE capella.installed_power_kw IS NOT NULL UNION ALL SELECT 'linda', linda.installed_power_kw FROM ferries_info WHERE linda.installed_power_kw IS NOT NULL UNION ALL SELECT 'sedna', sedna.installed_power_kw FROM ferries_info WHERE sedna.installed_power_kw IS NOT NULL UNION ALL SELECT 'ebba_brahe', ebba_brahe.installed_power_kw FROM ferries_info WHERE ebba_brahe.installed_power_kw IS NOT NULL ORDER BY installed_power_kw DESC LIMIT 1
