+++ backend/output/20250228_140748.txt
@@ -1,0 +1,16 @@
+QUERY:
+which ferry is the most powerful? based on ferries-info
+
+RESPONSE:
+The most powerful ferry is Merkurius with installed power of 1470 kW.
+
+SQL QUERIES:
+
+--- Query 1 ---
+SHOW TABLES
+
+--- Query 2 ---
+DESCRIBE ferries_info
+
+--- Query 3 ---
+SELECT 'fragancia' AS ferry, fragancia.installed_power_kw AS power FROM ferries_info UNION ALL SELECT 'jupiter', jupiter.installed_power_kw FROM ferries_info UNION ALL SELECT 'merkurius', merkurius.installed_power_kw FROM ferries_info UNION ALL SELECT 'nina', nina.installed_power_kw FROM ferries_info UNION ALL SELECT 'yxlan', yxlan.installed_power_kw FROM ferries_info UNION ALL SELECT 'marie', marie.installed_power_kw FROM ferries_info UNION ALL SELECT 'capella', capella.installed_power_kw FROM ferries_info UNION ALL SELECT 'linda', linda.installed_power_kw FROM ferries_info UNION ALL SELECT 'sedna', sedna.installed_power_kw FROM ferries_info UNION ALL SELECT 'ebba_brahe', ebba_brahe.installed_power_kw FROM ferries_info ORDER BY power DESC LIMIT 1
