+++ backend/output/20250228_145210.txt
@@ -1,0 +1,22 @@
+QUERY:
+which ferry is the most powerful? based on ferries-info
+
+RESPONSE:
+The most powerful ferry is **Merkurius** with an installed power of 1470 kW.
+
+SQL QUERIES:
+
+--- Query 1 ---
+SHOW TABLES File not found: /home/melke/lnu/2dv50e/lighthouse-sensor-bot/backend/data/ferries-info
+
+--- Query 2 ---
+CREATE TABLE IF NOT EXISTS 'ferries_info' AS SELECT * FROM '/home/melke/lnu/2dv50e/lighthouse-sensor-bot/backend/data/ferries-info' Using path from semantic model for table ferries_info: ferries.json
+
+--- Query 3 ---
+CREATE TABLE IF NOT EXISTS 'ferries_info' AS SELECT * FROM '/home/melke/lnu/2dv50e/lighthouse-sensor-bot/backend/data/ferries.json'
+
+--- Query 4 ---
+DESCRIBE ferries_info
+
+--- Query 5 ---
+SELECT 'fragancia' AS ferry_name, fragancia.installed_power_kw AS power_kw FROM ferries_info UNION ALL SELECT 'jupiter', jupiter.installed_power_kw FROM ferries_info UNION ALL SELECT 'merkurius', merkurius.installed_power_kw FROM ferries_info UNION ALL SELECT 'nina', nina.installed_power_kw FROM ferries_info UNION ALL SELECT 'yxlan', yxlan.installed_power_kw FROM ferries_info UNION ALL SELECT 'marie', marie.installed_power_kw FROM ferries_info UNION ALL SELECT 'capella', capella.installed_power_kw FROM ferries_info UNION ALL SELECT 'linda', linda.installed_power_kw FROM ferries_info UNION ALL SELECT 'sedna', sedna.installed_power_kw FROM ferries_info UNION ALL SELECT 'ebba_brahe', ebba_brahe.installed_power_kw FROM ferries_info ORDER BY power_kw DESC LIMIT 1
